[connector] Updated config handling and tests
